% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getAltData1}
\alias{getAltData1}
\alias{getAltData2}
\alias{setAltData1}
\alias{setAltData2}
\title{ALTREP API}
\usage{
getAltData1(x)

getAltData2(x)

setAltData1(x, value)

setAltData2(x, value)
}
\arguments{
\item{x}{An ALTREP object}

\item{value}{The data that will be set to an ALTREP object}
}
\value{
\code{getAltData1} : An R object in data1 slot of the ALTREP

\code{getAltData2} : An R object in data2 slot of the ALTREP

\code{setAltData1} : \code{TRUE} if the value has been set or
\code{FALSE} if the object is not an ALTREP

\code{setAltData2} : \code{TRUE} if the value has been set or
\code{FALSE} if the object is not an ALTREP
}
\description{
Get / Set data1 or data2 from an ALTREP object.
The function will return \code{NULL} if the object is
not an ALTREP. These are native ALTREP APIs, it is
not recommended to call these functions to modify an
ALTREP object.
}
\examples{
## A compact sequence is an ALTREP, 
## so ALTREP API can get the data that supports the compact sequence.
## The data is for developer only and not supposed to be used by users.
x=1:10
getAltData1(x)
getAltData2(x)

}
