% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALTREP-tools.R
\name{altrepMacro}
\alias{altrepMacro}
\alias{altrepSortStatus}
\alias{altrepNAStatus}
\title{ALTREP macro}
\description{
ALTREP macro
}
\details{
\code{altrepSortStatus} is a list containing all sortness status of an ALTREP object.
It can be used as the return value of \code{isSorted} function when defining
an ALTREP class. The sortness status can be unknown sortness,
known unsorted, ascending, descending, ascending with NA first,
descending with NA first.

Similarly, \code{altrepNAStatus} contains all possible NA status of an ALTREP object.
The current available status are unknow and no NA.
}
