% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altWrapper-tools.R
\name{isAltClassExist}
\alias{isAltClassExist}
\alias{isAltMethodExist}
\alias{getAltClassName}
\alias{getClassType}
\alias{getAltMethod}
\alias{getAltWrapperData}
\alias{setAltWrapperData}
\alias{deleteClass}
\alias{altClassStatus}
\title{AltWrapper APIs}
\usage{
isAltClassExist(className)

isAltMethodExist(className, methodName)

getAltClassName(x)

getClassType(className = NULL, x = NULL)

getAltMethod(className = NULL, methodName, x = NULL)

getAltWrapperData(x)

setAltWrapperData(x, value, duplicate = TRUE)

deleteClass(className, warning = TRUE)

altClassStatus(className = NULL, x = NULL)
}
\arguments{
\item{className}{Character, the name of an altWrapper class}

\item{methodName}{Character, the name of a function}

\item{x}{An altwrapper object}

\item{value}{The data of an altWrapper object}

\item{duplicate}{Whther duplicating the object(For the function \code{setAltWrapperData})}

\item{warning}{Logical, whether to give an warning if the class is not found.}
}
\value{
\code{isAltClassExist} : Logical

\code{isAltMethodExist} : Logical

\code{getAltClassName} : A character name

\code{getClassType} : A character indicating data type

\code{getAltMethod} : An R function or NULL value if the function is not found

\code{getAltWrapperData} : An R object

\code{setAltWrapperData} : The altWrapper object

\code{deleteClass} : No return value

\code{altClassStatus} : No return value
}
\description{
AltWrapper APIs
}
\details{
\code{isAltClassExist} : Whether an altWrapper class exist or not

\code{isAltMethodExist} : Whether an altWrapper method for
an altWrapper class exist or not. This function will return \code{FALSE}
if the altWrapper class does not exist.

\code{getAltClassName} : Get the altWrapper class name by
an altWrapper object

\code{getClassType} : Get the altWrapper class type by
a class name or an altWrapper object

\code{getAltMethod} : Get the altWrapper method by
a method name. Either \code{className} or \code{x} must be specified to
find the altWrapper class.

\code{getAltWrapperData} : Get the data for an altwrapper object

\code{setAltWrapperData} : Set the data of an altwrapper object, if
\code{duplicate} is TRUE, the function returns a duplication of the object.
Set \code{duplicate = FALSE} only when the new \code{value} does not change
the altWrapper object(e.g. caching on-dist data).

\code{deleteClass} : Delete an AltWrapper class

\code{altClassStatus} : Show the status of an altWrapper class
}
\examples{
## Define the ALTREP functions
length_func <- function(x) length(x)
get_ptr_func <- function(x,writeable) x


## Define the altWrapper class and its functions
setAltClass(className = "example", classType = "integer")
setAltMethod(className = "example", getLength = length_func)
setAltMethod(className = "example", getDataptr = get_ptr_func)

## Create an altWrapper object by providing the class name and data.
A = makeAltrep(className = "example", x = 1L:10L)
A

## Check the existance of an altWrapper class
isAltClassExist(className = "example")

## Check the existance of a function of an altWrapper class
isAltMethodExist(className = "example", methodName = "getLength")

## Get a function from an altWrapper class
getAltMethod(className = "example", methodName = "getLength")

## Get the class name by an altWrapper object
getAltClassName(A)

## Get the class type either by a class name or by an altWrapper object
getClassType(className = "example")
getClassType(x = A)

## Get the data of an altWrapper object
getAltWrapperData(x = A)

## Set the data of an altWrapper object
setAltWrapperData(x = A, value = 11L:20L)

## Show the status of an altWrapper class
## The class can be found either by its name or an altWrapper object
altClassStatus(className = "example")
altClassStatus(x = A)

}
