% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altWrapper-class.R
\docType{class}
\name{altNumeric-class}
\alias{altNumeric-class}
\alias{altWrapper-class}
\title{A convinent collection of altWrapper classes}
\description{
A convinent collection of altWrapper classes
}
