% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altWrapper-class.R
\docType{class}
\name{altWrapperClass}
\alias{altWrapperClass}
\alias{altWrapper-class}
\alias{altNumeric-class}
\alias{altRaw-class}
\alias{altLogical-class}
\alias{altInteger-class}
\alias{altDouble-class}
\title{The S3/S4 classes for ALTREP of type \code{raw}, \code{logical}, \code{integer} and \code{double}}
\description{
The S3/S4 classes for \code{raw}, \code{logical}, \code{integer} and \code{double} are \code{altRaw}, \code{altLogical},
\code{altInteger} and \code{altDouble} respectively. \code{altLogical}, \code{altInteger} and
\code{altDouble} inherit \code{altNumeric} class. \code{altRaw} and \code{altNumeric} inherit \code{altWrapper} class.
\code{altWrapper} is a subclass of \code{ANY} type.
}
