% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgSetting.R
\name{setAltWrapperOptions}
\alias{setAltWrapperOptions}
\alias{getAltWrapperOptions}
\title{AltWrapper Options Settings}
\usage{
setAltWrapperOptions(...)

getAltWrapperOptions(...)
}
\arguments{
\item{...}{a list of named arguments to change the package settings,
or a vector of character names to get the package settings. Currently the
only available setting is \code{redefineWarning}, which controls whether an warning
can be given when users try to redine a function or a class of an AltWrapper class.}
}
\value{
\code{setAltWrapperOptions}: No return value
\code{getAltWrapperOptions}: A list of package settings
}
\description{
Get and set the package settings
}
\examples{
## Get the default package settings
getAltWrapperOptions("redefineWarning")
## change the settings
setAltWrapperOptions(redefineWarning = FALSE)
## Get the settings again to see the change
getAltWrapperOptions("redefineWarning")

}
