% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altWrapper-export-tools.R
\name{duplicateObject}
\alias{duplicateObject}
\title{Duplicate an R object}
\usage{
duplicateObject(x, shallow)
}
\arguments{
\item{x}{The object that will be duplicated}

\item{shallow}{Logical, shallow duplicate or not}
}
\value{
An object of the same type as the input
}
\description{
Duplicate an R object, the argument shallow controls whether
to duplicate the container only or the entire object.
}
