% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altrep-export.R
\name{makeAltrep}
\alias{makeAltrep}
\title{Make an altWrapper object}
\usage{
makeAltrep(className, x, attributes = NULL)
}
\arguments{
\item{className}{Character, the name of an altWrapper class}

\item{x}{The data of the returned altWrapper object}

\item{attributes}{Named list, attributes that will be attached to the object}
}
\description{
This function make an altWrapper object from an altWrapper class
specified by the class name. The altWrapper class can
be set via \code{setAltClass} and \code{setAltMethod} functions.
}
\examples{
## Define the ALTREP functions
length_func<-function(x) length(x)
get_ptr_func<-function(x,writeable) x

## Define a function to compute variance for the altWrapper object.
var_func<-function(x, y = NULL, na.rm = FALSE, use) return()

## Define the altWrapper class and its functions
setAltClass(className = "example", classType = "integer")
setAltMethod(className = "example", getLength = length_func)
setAltMethod(className = "example", getDataptr = get_ptr_func)

## Create an altWrapper object by providing the class name and data.
A=makeAltrep(className = "example", x = 1L:10L)
A


}
