% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altWrapper-main.R
\name{newAltrep}
\alias{newAltrep}
\title{Make an altWrapper object}
\usage{
newAltrep(className, x, attributes = list(), S3Class = FALSE,
  S4Class = FALSE)
}
\arguments{
\item{className}{Character, the name of an altWrapper class}

\item{x}{The data of the returned altWrapper object}

\item{attributes}{Named list, attributes that will be attached to the object. The
attribute \code{class} will be ignored}

\item{S3Class}{Logical or character vector, whether the return value is of an
altWrapper S3 class. If the argument is a character vector, it will overwrite the
default settings and resulting in an S3 object with the class type specified in the
character vector.}

\item{S4Class}{Logical or character, whether the return value is of
an AltWrapper S4 class. If the argument is a character, it must be a class name and
the return value is an S4 object specified by the class name.
An error will be returned if both \code{S3Class} and \code{S4Class} are defined.}
}
\value{
An ALTREP vector
}
\description{
This function make an altWrapper object from an altWrapper class
specified by the class name. The altWrapper class can
be set via \code{setAltClass} and \code{setAltMethod} functions.
}
\examples{
## Define the ALTREP functions
length_func <- function(x) length(x)
get_ptr_func <- function(x,writeable) x


## Define the altWrapper class and its functions
setAltClass(className = "example", classType = "integer")
setAltMethod(className = "example", getLength = length_func)
setAltMethod(className = "example", getDataptr = get_ptr_func)

## Create an altWrapper object by providing the class name and data.
A <- newAltrep(className = "example", x = 1L:10L)
A


}
