% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altWrapper-class.R
\name{altWrapperPrint}
\alias{altWrapperPrint}
\alias{print.altRaw}
\alias{print.altInteger}
\alias{print.altLogical}
\alias{print.altReal}
\alias{printAltWrapper}
\title{Print altWrapper vector values}
\usage{
\method{print}{altRaw}(x, ...)

\method{print}{altInteger}(x, ...)

\method{print}{altLogical}(x, ...)

\method{print}{altReal}(x, ...)

printAltWrapper(x, ...)
}
\arguments{
\item{x}{An altWrapper object}

\item{...}{No effect, for compatibility only}
}
\value{
The argument \code{x}
}
\description{
This function is a complement of the print function. It is able to
print an ALTREP object made by AltWrapper package. In case that the
data pointer is not available, the function will look at
ALTREP APIs in the order of \code{getRegion}, \code{getSubset},
\code{getElement} to search for the available way to collect the data.
}
\examples{
length_func <- function(x) length(x)
element_function <- function(x,i) x[i]

## Define the altWrapper class and its functions
setAltClass(className = "example", classType = "integer")
setAltMethod(className = "example", getLength = length_func)
setAltMethod(className = "example", getElement = element_function)


## Create an altWrapper object by providing the class name and data.
A=makeAltrep(className = "example", x = 1L:10L)

## The default print function does not wrok since it uses
## data pointer only.
\dontrun{
A
}

## The package provide an alternative print function
printAltWrapper(A)

}
