% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altWrapper-tools.R
\name{removeWrapper}
\alias{removeWrapper}
\title{Remove the wrapper created by R}
\usage{
removeWrapper(x)
}
\arguments{
\item{x}{An altWrapper object}
}
\value{
An altWrapper object
}
\description{
This function can only be used on an altWrapper object.
The wrapper is an ALTREP object and created by R to automatically memorize
some statistics(e.g. sum). it can be silently made in some R functions and behaves exactly
the same as the object it wraps. This function can be called
in order to make sure the altWrapper variable \code{x} is not wrapped by a wrapper,
}
\examples{
## Define an altWrapper class
length_func<-function(x) length(x)
setAltClass(className = "example", classType = "integer")
setAltMethod(className = "example", getLength = length_func)
A <- newAltrep(className = "example", x = 1L:10L)

## Since A is a new object and does not have a wrapper,
## calling `removeWrapper` does not have any effect.
A <- removeWrapper(A)
}
