% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altWrapper-class.R
\docType{class}
\name{altwrapper-S4class}
\alias{altwrapper-S4class}
\alias{altRaw-class}
\alias{altLogical-class}
\alias{altInteger-class}
\alias{altReal-class}
\title{The S4 class for ALTREP of type raw, logical, integer and real}
\description{
The S4 class for ALTREP of type raw, logical, integer and real
}
