% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altWrapper-main.R
\name{setAltClass}
\alias{setAltClass}
\title{Define an altWrapper class}
\usage{
setAltClass(className, classType = c("raw", "logical", "integer",
  "double"))
}
\arguments{
\item{className}{Character, the name of the class}

\item{classType}{Character, the type of the altWrapper}
}
\value{
No return value
}
\description{
Define an altWrapper class by specifying its name and data type
}
\examples{
## Define the ALTREP functions
length_func <- function(x) length(x)
get_ptr_func <- function(x,writeable) x


## Define the altWrapper class and its functions
setAltClass(className = "example", classType = "integer")
setAltMethod(className = "example", getLength = length_func)
setAltMethod(className = "example", getDataptr = get_ptr_func)

## Create an altWrapper object by providing the class name and data.
A <- newAltrep(className = "example", x = 1L:10L)
A


}
