% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altrep-export.R
\name{setAltData}
\alias{setAltData}
\title{Set the data for an altwrapper object in ALTREP functions}
\usage{
setAltData(x)
}
\description{
Set the data for an altwrapper object,
The function is only allowed to be called in an ALTREP API.
Please refer to \code{?setAltMethod} to find a full list of the ALTREP APIs.
}
