% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altWrapper-tools.R
\name{setAltSelfData}
\alias{setAltSelfData}
\title{Set the data for an altWrapper object in ALTREP functions}
\usage{
setAltSelfData(x)
}
\arguments{
\item{x}{The data of an altWrapper object}
}
\value{
no return value
}
\description{
Set the data for an altWrapper object inside an ALTREP function,
The function is only allowed to be called in an ALTREP API.
It is designed for caching data for an altWrapper object(eg. sum value,
a block of on-disk data). It CANNOT be used to change the value of the
altWrapper object.
Please refer to \code{?setAltMethod} to find a full list of the ALTREP APIs.
}
\examples{
## The data is a list
data <- list(data=runif(10))

## Define the ALTREP functions
length_func <- function(x) length(x$data)
get_ptr_func <- function(x,writeable) x$data

## Define a sum function,
## after the fisrt computation,
## store the result in the list.
## Call `setAltSelfData` to set the data
sum_func <- function(x, na.rm) {
   if(is.null(x$sum)){
     message("computing sum")
     x$sum <- sum(x$data)
     setAltSelfData(x)
     return(x$sum)
   }else{
     message("reading sum result")
     return(x$sum)
   }
}

## Define the altWrapper class and its functions
setAltClass(className = "example", classType = "double")
setAltMethod(className = "example", getLength = length_func)
setAltMethod(className = "example", getDataptr = get_ptr_func)
setAltMethod(className = "example", sum = sum_func)

A <- newAltrep(className = "example", x = data)

#First sum call
sum(A)
#Second sum call
sum(A)
}
