% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALTREP-tools.R, R/altWrapper-export-tools.R
\name{is.altrep}
\alias{is.altrep}
\alias{is.altWrapper}
\title{Check the type of an object}
\usage{
is.altrep(x)

is.altWrapper(x)
}
\arguments{
\item{x}{An object}
}
\value{
Logical value indicating whether the object is an ALTREP

Logical value indicating whether the object is an altWrapper
}
\description{
Check the type of an object
}
\details{
\code{is.altrep} : Is an object ALTREP?

\code{is.altWrapper} : Is an object altWrapper?
}
\examples{
## An R vector is not ALTREP nor altWrapper.
x=runif(10)
is.altrep(x)
is.altWrapper(x)

## A compact sequence is ALTREP but not altWrapper.
x=1:10
is.altrep(x)
is.altWrapper(x)

## An altWrapper object is ALTREP and altWrapper
length_func<-function(x) length(x)
setAltClass(className = "example", classType = "integer")
setAltMethod(className = "example", getLength = length_func)
A=makeAltrep(className = "example", x = 1L:10L)
is.altrep(x)
is.altWrapper(A)
}
