% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altrep-export.R
\name{isAltClassExist}
\alias{isAltClassExist}
\alias{isAltMethodExist}
\alias{getAltClassName}
\alias{getClassType}
\alias{getAltMethod}
\alias{getAltWrapperData}
\alias{setAltWrapperData}
\alias{removeClass}
\alias{altClassStatus}
\title{AltWrapper APIs}
\usage{
isAltClassExist(className)

isAltMethodExist(className, methodName)

getAltClassName(x)

getClassType(className = NULL, x = NULL)

getAltMethod(className = NULL, methodName, x = NULL)

getAltWrapperData(x)

setAltWrapperData(x, value)

removeClass(className)

altClassStatus(className = NULL, x = NULL)
}
\arguments{
\item{className}{Character, the name of an altWrapper class}

\item{methodName}{Character, the name of a function}

\item{x}{an altwrapper object}
}
\description{
AltWrapper APIs
}
\details{
\code{isAltClassExist} : Whether an altWrapper class exist or not

\code{isAltMethodExist} : Whether an altWrapper method for
an altWrapper class exist or not. This function will return \code{FALSE}
if the altWrapper class does not exist.

\code{getAltClassName} : Get the altWrapper class name by
an altWrapper object

\code{getClassType} : Get the altWrapper class type by
a class name or an altWrapper object

\code{getAltMethod} : Get the altWrapper method by
a method name. Either \code{className} or \code{x} must be specified to
find the altWrapper class. The method can be a wrapper method of ALTREP API
or a method attached to the altWrapper class.

\code{getAltWrapperData} : Get the data for an altwrapper object

\code{setAltWrapperData} : Set the data for an altwrapper object

\code{removeClass} : Remove an AltWrapper class

\code{altClassStatus} : Show the status of an altWrapper class
}
\examples{
aaa

}
