% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altWrapper-tools.R
\name{isAltClassDefined}
\alias{isAltClassDefined}
\alias{isAltMethodDefined}
\alias{getAltClassName}
\alias{getAltClassType}
\alias{getAltMethod}
\alias{getAltWrapperData}
\alias{setAltWrapperData}
\alias{deleteAltClass}
\alias{showAltClass}
\title{AltWrapper APIs}
\usage{
isAltClassDefined(className)

isAltMethodDefined(className, methodName)

getAltClassName(x)

getAltClassType(className = NULL, x = NULL)

getAltMethod(className = NULL, methodName, x = NULL)

getAltWrapperData(x)

setAltWrapperData(x, value, duplicate = TRUE)

deleteAltClass(className, warning = TRUE)

showAltClass(className = NULL, x = NULL)
}
\arguments{
\item{className}{Character, the name of an altWrapper class}

\item{methodName}{Character, the name of a function}

\item{x}{An altWrapper object}

\item{value}{The data of an altWrapper object}

\item{duplicate}{Whther duplicating the object(For the function \code{setAltWrapperData})}

\item{warning}{Logical, whether to give an warning if the class is not found.}
}
\value{
\code{isAltClassDefined} : Logical

\code{isAltMethodDefined} : Logical

\code{getAltClassName} : A character name

\code{getAltClassType} : A character indicating data type

\code{getAltMethod} : An R function or NULL value if the function is not found

\code{getAltWrapperData} : An R object

\code{setAltWrapperData} : The altWrapper object

\code{deleteAltClass} : No return value

\code{showAltClass} : No return value
}
\description{
AltWrapper APIs
}
\details{
\code{isAltClassDefined} : Whether an altWrapper class exist or not

\code{isAltMethodDefined} : Whether an altWrapper method for
an altWrapper class exist or not. This function will return \code{FALSE}
if it cannot find the altWrapper class.

\code{getAltClassName} : Get the altWrapper class name by
an altWrapper object

\code{getAltClassType} : Get the altWrapper class type by
a class name or an altWrapper object

\code{getAltMethod} : Get the altWrapper method by
a method name. Either \code{className} or \code{x} must be specified to
find the altWrapper class.

\code{getAltWrapperData} : Get the data for an altWrapper object

\code{setAltWrapperData} : Set the data of an altWrapper object, if
\code{duplicate} is TRUE, the function returns a duplication of the object.
Set \code{duplicate = FALSE} only when the new \code{value} does not change
the altWrapper object(e.g. caching on-dist data).

\code{deleteAltClass} : Delete an AltWrapper class

\code{showAltClass} : Show the status of an altWrapper class
}
\examples{
## Define the ALTREP functions
length_func <- function(x) length(x)
get_ptr_func <- function(x,writeable) x


## Define the altWrapper class and its functions
setAltClass(className = "example", classType = "integer")
setAltMethod(className = "example", getLength = length_func)
setAltMethod(className = "example", getDataptr = get_ptr_func)

## Create an altWrapper object by providing the class name and data.
A <- newAltrep(className = "example", x = 1L:10L)
A

## Check the existance of an altWrapper class
isAltClassDefined(className = "example")

## Check the existance of a function of an altWrapper class
isAltMethodDefined(className = "example", methodName = "getLength")

## Get a function from an altWrapper class
getAltMethod(className = "example", methodName = "getLength")

## Get the class name by an altWrapper object
getAltClassName(A)

## Get the class type either by a class name or by an altWrapper object
getAltClassType(className = "example")
getAltClassType(x = A)

## Get the data of an altWrapper object
getAltWrapperData(x = A)

## Set the data of an altWrapper object
setAltWrapperData(x = A, value = 11L:20L)

## Show the status of an altWrapper class
## The class can be found either by its name or an altWrapper object
showAltClass(className = "example")
showAltClass(x = A)

}
