% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altrep-export.R
\name{attachAltMethod}
\alias{attachAltMethod}
\title{Attach ALTREP method}
\usage{
attachAltMethod(className, ...)
}
\arguments{
\item{className}{Character, the name of an altWrapper class}

\item{...}{Named arguments. The functions that you want to attach to the class.}
}
\description{
Attach a function to an altwrapper class. The function will
not be called by R automatically. It is designed to mimic
the object-oriented feature for an altwrapper class.
}
