% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{removeWrapper}
\alias{removeWrapper}
\title{Remove the wrapper created by R}
\usage{
removeWrapper(x)
}
\arguments{
\item{x}{An altWrapper object}
}
\value{
An altWrapper object
}
\description{
This function can only be used with an altWrapper object. It
will remove the wrapper created by R.
}
\examples{
## Define an altWrapper class
length_func<-function(x) length(x)
setAltClass(className = "example", classType = "integer")
setAltMethod(className = "example", getLength = length_func)
A=makeAltrep(className = "example", x = 1L:10L)

## Since A is a new object and does not have a wrapper,
## calling `removeWrapper` does not have any effect.
A=removeWrapper(A)
}
